+++ src/db.js
@@ -1,11 +1,11 @@
 'use strict';
 
 const pgp = require('pg-promise')();
-const user = 'guseruser';
+const user = process.env.PG_USER;
 const password = process.env.PG_PASS;
-const host = 'localhost';
-const port = '5432';
-const database = 'guser';
+const host = process.env.PG_HOST;
+const port = process.env.PG_PORT;
+const database = process.env.PG_DB;
 const db = pgp('postgres://' + user + ':' + password + '@' + host + ':' + port + '/' + database);
 
 module.exports = db;